Add files via upload
In this project we are utilizing a data file consisting of all the words along with their synonyms as the base file. we are fetching an input word from the user and then checking its validity and conveying its meaning. Here this project is unbiased by case sensitive issues, minute misspelling issues and so on.. by the usage of several hybrid methods and importing strategies.